# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SamaCloud",
        "version": "0.1.0",
        "description": "API de contr\u00f4le de SamaCloud : comptes, jetons, projets, d\u00e9ploiements, commandes et \u00e9change.\n\nCette sp\u00e9cification est une **proposition d'API pour Datacloud**, \u00e9crite pour le concours CADEV. Ce n'est pas l'API officielle de Systalink.\n\nToutes les r\u00e9ponses de succ\u00e8s sont envelopp\u00e9es dans `data`. Toutes les erreurs ont la forme `{ code, message, details? }`, avec un message en fran\u00e7ais."
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur qui h\u00e9berge cette documentation"
        }
    ],
...
    "security": [
        {
            "jeton": []
        }
    ],
    "paths": {
...
        "/jetons-ia/{id}": {
            "delete": {
                "operationId": "agent-tokens.destroy",
                "description": "Effet imm\u00e9diat. Un jeton d'un autre compte, ou un jeton de session, est \u00ab introuvable \u00bb.",
                "summary": "R\u00e9voquer un jeton IA",
                "tags": [
                    "Jetons IA"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/questions/{id}": {
            "get": {
                "operationId": "questions.show",
                "description": "Lecture publique. Les 20 premi\u00e8res r\u00e9ponses sont jointes ; les suivantes se lisent avec\nGET /api/questions/{id}/reponses. Avec un jeton, \u00ab vote_par_moi \u00bb indique les r\u00e9ponses\npour lesquelles le compte a d\u00e9j\u00e0 vot\u00e9 \u00ab Utile \u00bb.",
                "summary": "Voir une question et ses r\u00e9ponses",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la question.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`QuestionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/QuestionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                },
                "security": []
            },
            "patch": {
                "operationId": "questions.update",
                "description": "R\u00e9serv\u00e9 \u00e0 l'auteur. Seuls les champs envoy\u00e9s sont modifi\u00e9s.",
                "summary": "Modifier sa question",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la question.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateQuestionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`QuestionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/QuestionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "- `validation_echouee` \u2014 Donn\u00e9es invalides ; \u00ab details \u00bb liste chaque champ fautif.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "questions.destroy",
                "description": "R\u00e9serv\u00e9 \u00e0 l'auteur, ou \u00e0 un administrateur pour la mod\u00e9ration. La question et ses r\u00e9ponses\nne sont plus visibles.",
                "summary": "Supprimer sa question",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la question.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/reponses/{id}": {
            "patch": {
                "operationId": "answers.update",
                "description": "R\u00e9serv\u00e9 \u00e0 l'auteur de la r\u00e9ponse.",
                "summary": "Modifier sa r\u00e9ponse",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la r\u00e9ponse.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AnswerResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AnswerResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "- `validation_echouee` \u2014 Donn\u00e9es invalides ; \u00ab details \u00bb liste chaque champ fautif.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "answers.destroy",
                "description": "R\u00e9serv\u00e9 \u00e0 l'auteur, ou \u00e0 un administrateur pour la mod\u00e9ration. Si c'\u00e9tait la r\u00e9ponse accept\u00e9e,\nla question redevient \u00ab non r\u00e9solue \u00bb. Les points qu'elle avait rapport\u00e9s sont repris.",
                "summary": "Supprimer sa r\u00e9ponse",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la r\u00e9ponse.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/questions/{id}/signalements": {
            "post": {
                "operationId": "reports.question",
                "description": "Pr\u00e9vient les administrateurs qu'une question pose probl\u00e8me. Signaler deux fois\nla m\u00eame question ne cr\u00e9e pas de doublon.",
                "summary": "Signaler une question",
                "tags": [
                    "Communaut\u00e9 \u2014 questions-r\u00e9ponses"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Identifiant de la question.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`ReportResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ReportResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "- `non_authentifie` \u2014 Jeton absent, expir\u00e9 ou r\u00e9voqu\u00e9.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "- `compte_suspendu` \u2014 Le compte a \u00e9t\u00e9 suspendu par un administrateur.\n- `capacite_manquante` \u2014 Le jeton n'a pas la capacit\u00e9 exig\u00e9e par cette route.\n- `signalement_de_son_contenu` \u2014 On ne signale pas son propre contenu.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "- `introuvable` \u2014 La ressource n'existe pas ou n'appartient pas \u00e0 ce compte.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "- `validation_echouee` \u2014 Donn\u00e9es invalides ; \u00ab details \u00bb liste chaque champ fautif.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "- `trop_de_requetes` \u2014 Limite de d\u00e9bit atteinte ; voir \u00ab details.reessayer_dans_secondes \u00bb.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erreur"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jeton": {
                "type": "http",
                "description": "Jeton de session (`sc_sess_\u2026`) ou jeton d'agent IA (`sc_live_\u2026`, `sc_test_\u2026`).",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "AnswerRequest": {
                "type": "object",
                "description": "Contenu d'une r\u00e9ponse (cr\u00e9ation ou modification).",
                "properties": {
                    "contenu": {
                        "type": "string",
                        "description": "Markdown, avec blocs de code.",
                        "minLength": 10,
                        "maxLength": 20000
                    }
                },
                "required": [
                    "contenu"
                ],
                "title": "AnswerRequest"
            },
...
            "Erreur": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Identifiant stable de l'erreur, \u00e0 lire par les programmes.",
                        "examples": [
                            "quota_atteint"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "description": "Explication en fran\u00e7ais, affichable telle quelle."
                    },
                    "details": {
                        "description": "Pr\u00e9cisions \u00e9ventuelles. Pour une erreur de validation : liste de { champ, message }."
                    }
                },
                "required": [
                    "code",
                    "message"
                ],
                "title": "Erreur"
            },
...
            "QuestionResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "titre": {
                        "type": "string"
                    },
                    "contenu": {
                        "type": "string",
                        "description": "Markdown brut : c'est \u00e0 l'affichage de le mettre en forme et de le nettoyer."
                    },
                    "technologies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TechnologyResource"
                        }
                    },
                    "auteur": {
                        "$ref": "#/components/schemas/AuthorResource"
                    },
                    "nb_reponses": {
                        "type": "integer"
                    },
                    "resolue": {
                        "type": "boolean"
                    },
                    "reponse_acceptee_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "reponses": {
                        "type": "array",
                        "description": "Les 20 premi\u00e8res : la r\u00e9ponse accept\u00e9e d'abord, puis les plus utiles, puis les plus anciennes.\nAu-del\u00e0 (voir \u00ab nb_reponses \u00bb), lire GET /api/questions/{id}/reponses?page=2.",
                        "items": {
                            "$ref": "#/components/schemas/AnswerResource"
                        }
                    },
                    "cree_le": {
                        "type": "string"
                    },
                    "modifie_le": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "titre",
                    "contenu",
                    "technologies",
                    "auteur",
                    "nb_reponses",
                    "resolue",
                    "reponse_acceptee_id",
                    "reponses",
                    "cree_le",
                    "modifie_le"
                ],
                "title": "QuestionResource"
            },
...
            "UpdateQuestionRequest": {
                "type": "object",
                "description": "Modification d'une question. Seuls les champs envoy\u00e9s sont modifi\u00e9s.",
                "properties": {
                    "titre": {
                        "type": "string",
                        "minLength": 10,
                        "maxLength": 150
                    },
                    "contenu": {
                        "type": "string",
                        "minLength": 20,
                        "maxLength": 20000
                    },
                    "technologies": {
                        "type": "array",
                        "description": "Liste compl\u00e8te des technologies \u00e0 garder.",
                        "items": {
                            "type": "string"
                        },
                        "minItems": 1,
                        "maxItems": 5,
                        "uniqueItems": true
                    }
                },
                "title": "UpdateQuestionRequest"
            },
...
        }
    }
}